fix: Remove context towards models creation
