+++ packages/graphback-codegen-resolvers/src/ResolverGeneratorPlugin.ts
@@ -3,7 +3,7 @@ import { OutputFileSystem } from './GeneratorModel';
 import { GeneratorResolversFormat } from './GeneratorResolversFormat';
 import { generateCRUDResolversFunctions } from './templates/createResolvers';
 import { createResolverTemplate } from './templates/resolverWrapper';
-import { createRuntimeFile } from './templates/runtimeTemplate';
+import { createModelsFile } from './templates/modelTemplate';
 import { writeResolvers } from './writeResolvers';
 
 export interface ResolverGeneratorPluginConfig {
@@ -80,7 +80,7 @@ export class ResolverGeneratorPlugin extends GraphbackPlugin {
 
         const generatedResolversFunctions = generateCRUDResolversFunctions(models);
         const generatedResolverFile = this.createGeneratedResolversFile(generatedResolversFunctions, this.pluginConfig);
-        const contextFile = createRuntimeFile(models, this.pluginConfig);
+        const contextFile = createModelsFile(models, this.pluginConfig);
 
         return {
             resolvers: generatedResolverFile,
